backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.25.0
google-generativeai==0.8.0
google-api-python-client==2.127.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
firebase-admin==6.5.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
